Replace `Int` fields with `Long` in other models

+++ app/src/main/java/io/github/frankolt/githubexplorer/data/http/github/models/OwnerResponse.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class OwnerResponse(
     @Json(name = "login") val login: String? = null,
-    @Json(name = "id") val id: Int? = null,
+    @Json(name = "id") val id: Long? = null,
     @Json(name = "node_id") val nodeId: String? = null,
     @Json(name = "avatar_url") val avatarUrl: String? = null,
     @Json(name = "gravatar_url") val gravatarId: String? = null,
